refactor restaurant menu to food menu and add pagination

+++ src/modules/food-menu/food-menu.service.ts
@@ -1,29 +1,30 @@
 import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common'
 import { RestaurantEntity } from '../restaurant/persistence/Restaurant.entity'
-import { CreateRestaurantMenuRequestDto } from './dto/create-restaurant-menu-request.dto'
-import { UpdateRestaurantMenuRequestDto } from './dto/update-restaurant-menu-request.dto'
-import { RestaurantMenuDaoService } from './persistence/restaurant-menu.dao.service'
-import { RestaurantMenuMapper } from './domain/restaurant-menu.mapper'
+import { CreateFoodMenuRequestDto } from './dto/create-food-menu-request.dto'
+import { UpdateFoodMenuRequestDto } from './dto/update-food-menu-request.dto'
+import { FoodMenuDaoService } from './persistence/food-menu.dao.service'
+import { FoodMenuMapper } from './domain/food-menu.mapper'
 import { EntityManager } from '@mikro-orm/postgresql'
 import { MenuCategoryMapper } from './domain/menu-category.mapper'
-import { fstat, unlink } from 'fs'
+import { unlink } from 'fs'
+import { GetAllFoodMenuQueryDto } from './dto/get-all-food-menu-query.dto'
 
 @Injectable()
-export class RestaurantMenuService {
+export class FoodMenuService {
 	constructor(
-		private readonly restaurantMenuDaoService: RestaurantMenuDaoService,
+		private readonly foodMenuDaoService: FoodMenuDaoService,
 		private readonly em: EntityManager
 	) {}
 
 	async createMenu(
 		restaurant: RestaurantEntity | null,
 		menuPicture: Express.Multer.File,
-		data: CreateRestaurantMenuRequestDto
+		data: CreateFoodMenuRequestDto
 	) {
 		if (!restaurant) {
 			throw new BadRequestException('Unauthorized Request')
 		}
-		const menu = await this.restaurantMenuDaoService.createMenu({
+		const menu = await this.foodMenuDaoService.createMenu({
 			...data,
 			menuPictureUrl: `files/menu/${menuPicture.filename}`,
 			restaurant: restaurant
@@ -31,7 +32,7 @@ export class RestaurantMenuService {
 
 		await this.em.flush()
 
-		const menuMapped = RestaurantMenuMapper.toDomain(menu)
+		const menuMapped = FoodMenuMapper.toDomain(menu)
 		return {
 			menu: menuMapped
 		}
@@ -40,13 +41,13 @@ export class RestaurantMenuService {
 	async updateMenu(
 		restaurant: RestaurantEntity | null,
 		menuId: string,
-		data: UpdateRestaurantMenuRequestDto,
+		data: UpdateFoodMenuRequestDto,
 		menuPicture?: Express.Multer.File
 	) {
 		if (!restaurant) {
 			throw new BadRequestException('Unauthorized Request')
 		}
-		const menu = await this.restaurantMenuDaoService.findMenuById(menuId)
+		const menu = await this.foodMenuDaoService.findMenuById(menuId)
 		if (!menu) {
 			throw new NotFoundException('Menu not found')
 		}
@@ -55,21 +56,23 @@ export class RestaurantMenuService {
 		}
 
 		if (menuPicture) {
-			unlink(`src/${menu.pictureUrl}`, (err) => {
-				if (err) throw new BadRequestException('Error deleting menu picture')
-				console.log('Menu picture deleted')
-			})
+			if (!menu.pictureUrl.includes('dummy')) {
+				unlink(`src/${menu.pictureUrl}`, (err) => {
+					if (err) throw new BadRequestException('Error deleting menu picture')
+					console.log('Menu picture deleted')
+				})
+			}
 		}
 
-		await this.restaurantMenuDaoService.updateMenu(menu, {
+		await this.foodMenuDaoService.updateMenu(menu, {
 			...data,
 			menuPictureUrl: menuPicture ? `files/menu/${menuPicture.filename}` : menu.pictureUrl
 		})
 
 		await this.em.flush()
 
-		const updatedMenu = await this.restaurantMenuDaoService.findMenuById(menuId)
-		const menuMapped = RestaurantMenuMapper.toDomain(updatedMenu)
+		const updatedMenu = await this.foodMenuDaoService.findMenuById(menuId)
+		const menuMapped = FoodMenuMapper.toDomain(updatedMenu)
 		return {
 			menu: menuMapped
 		}
@@ -84,7 +87,7 @@ export class RestaurantMenuService {
 			throw new BadRequestException('Unauthorized Request')
 		}
 
-		const menu = await this.restaurantMenuDaoService.findMenuById(menuId)
+		const menu = await this.foodMenuDaoService.findMenuById(menuId)
 
 		if (!menu) {
 			throw new NotFoundException('Menu not found')
@@ -94,12 +97,12 @@ export class RestaurantMenuService {
 		}
 		console.log(isAvailable)
 
-		this.restaurantMenuDaoService.updateMenuStockAvailablity(menu, isAvailable)
+		this.foodMenuDaoService.updateMenuStockAvailablity(menu, isAvailable)
 		await this.em.flush()
-		const updatedMenu = await this.restaurantMenuDaoService.findMenuById(menuId)
+		const updatedMenu = await this.foodMenuDaoService.findMenuById(menuId)
 
 		return {
-			updatedMenu: RestaurantMenuMapper.toDomain(updatedMenu)
+			updatedMenu: FoodMenuMapper.toDomain(updatedMenu)
 		}
 	}
 
@@ -107,25 +110,24 @@ export class RestaurantMenuService {
 		if (!restaurant) {
 			throw new BadRequestException('Unauthorized Request')
 		}
-		const menu = await this.restaurantMenuDaoService.findMenuById(menuId)
+		const menu = await this.foodMenuDaoService.findMenuById(menuId)
 		if (!menu) {
 			throw new NotFoundException('Menu not found')
 		}
 		if (menu.restaurant.id !== restaurant.id) {
 			throw new BadRequestException('Unauthorized Request')
 		}
-		this.restaurantMenuDaoService.removeMenu(menu)
+		this.foodMenuDaoService.removeMenu(menu)
 		await this.em.flush()
 		return {
 			message: 'Menu removed successfully'
 		}
-
 	}
 
-	async findAllMenusByRestaurantId(restaurantId: string) {
-		const menus = await this.restaurantMenuDaoService.findMenusByRestaurantId(restaurantId)
+	async findAllMenusByRestaurantId(restaurantId: string, query: GetAllFoodMenuQueryDto) {
+		const menus = await this.foodMenuDaoService.findMenusByRestaurantId(restaurantId, query)
 		const menusMapped = menus
-			.map((menu) => RestaurantMenuMapper.toDomain(menu))
+			.map((menu) => FoodMenuMapper.toDomain(menu))
 			.filter((menu) => menu !== null)
 		return {
 			menus: menusMapped
@@ -133,7 +135,7 @@ export class RestaurantMenuService {
 	}
 
 	async getAllMenuCategories() {
-		const categories = await this.restaurantMenuDaoService.getAllMenuCategories()
+		const categories = await this.foodMenuDaoService.getAllMenuCategories()
 		return {
 			categories: categories.map((category) => MenuCategoryMapper.toDomain(category))
 		}
